プロファイルロード時にクリップに値をディープコピーするように変更
